+++ packages/hooks-registry/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackwright/hooks-registry
 
+## 0.1.4-alpha.1
+
+### Patch Changes
+
+- Updated dependencies [3bac6b8]
+  - @stackwright/types@1.8.0-alpha.1
+
+## 0.1.4-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [803e6ea]
+  - @stackwright/types@1.7.1-alpha.0
+
 ## 0.1.3
 
 ### Patch Changes